assorted changes
* send partial product info in the item json file
* refactor checkout page
* Fix right rounded corners in the form-control in login and checkout
  page(this issue occurs when invalid check is added -> add
  rounded-right to temporarily fix this)
* Fixes for router and axios
* Fix Header gap in the Dashboard layout

+++ backend/api.py
@@ -572,7 +572,7 @@ order_request_parse.add_argument("name", type=str, required=True)
 order_request_parse.add_argument("address", type=str, required=True)
 order_request_parse.add_argument("phone_number", type=int, required=True)
 order_request_parse.add_argument("items",type=str,action='append',location='json', required=True)
-order_request_parse.add_argument("total_amount", type=float, required=True)
+order_request_parse.add_argument("total_amount", type=int, required=True)
 order_request_parse.add_argument("delivery_date", type=valid_date, required=True)
 # order_request_parse.add_argument("user_id", type=int)
 class OrderAPI(Resource):
